requirements.txt
customtkinter==5.2.2
Pillow>=10.0.0
PyPDF2>=3.0.0
tkinterdnd2>=0.3.0
packaging>=23.0
pytest>=7.0.0
ruff>=0.1.0
